Moved localhost_apiserver CloudExport stub server into cloudexport directory
